src/php-backend/tools/AI_tool/requirements.txt
fastapi
uvicorn
python-dotenv
openai
PyPDF2
python-docx
pandas
fpdf2
python-multipart
Requests
pytesseract
pdf2image
Pillow